Some logic around consecutive failures

+++ src/main.rs
@@ -13,6 +13,7 @@ const IPS_TO_PING: [IpAddr; 3] = [
 
 #[tokio::main]
 async fn main() {
+    let mut times_in_a_row = 0;
     loop {
         let mut pings = vec![];
         for ip in IPS_TO_PING {
@@ -40,11 +41,20 @@ async fn main() {
         }
 
         if network_down_errors == IPS_TO_PING.len() {
-            // Reboots the computer
-            if let Err(e) = system_shutdown::force_reboot() {
-                eprintln!("Error shutting down PC: {:?}", e);
-                Command::new("reboot").args(&["-r", "now"]).spawn().unwrap();
-            };
+            // Add one to the times this has happened in a row
+            times_in_a_row += 1;
+
+            // If we're down for a full minute (6 times * 10 seconds sleep timer), then reboot
+            if times_in_a_row >= 6 {
+                // Reboots the computer
+                if let Err(e) = system_shutdown::force_reboot() {
+                    eprintln!("Error shutting down PC: {:?}", e);
+                    Command::new("reboot").args(&["-r", "now"]).spawn().unwrap();
+                };
+            }
+        } else {
+            // If network is up, clear the consecutive failed count
+            times_in_a_row = 0;
         }
 
         tokio::time::sleep(Duration::from_secs(10)).await;
